huge sweeping restructure of the project.
even though we're ignoring standard python conventions by using camel case, I feel like standard python organization shouldn't be much of an issue. I've moved everything into packages and made a base level package. I've moved several loose functions to a helpers file and the custom classes to the Core package. I made an override to the load_extension method to make this move not be a hindrance. this does present some issues if we plan on there being external cogs, but that is addressable as well.
