+++ RyuZU/Admin/__init__.py
@@ -1,6 +1,5 @@
+from RyuZU.Core import Cog
 from discord.ext import commands
-
-from Cog import Cog
 
 
 class Admin(Cog):
