+++ RyuZU/MusicPlayer/__init__.py
@@ -3,9 +3,8 @@ import math
 import platform
 
 import discord
+from RyuZU.Core import Cog
 from discord.ext import commands
-
-from Cog import Cog
 
 if not discord.opus.is_loaded() and platform.system() == "Windows":
     discord.opus.load_opus('opus')
