+++ RyuZU/Quotes/__init__.py
@@ -1,9 +1,8 @@
 import random
 
+from RyuZU.Core import Cog
 from discord.ext import commands
 from tinydb import Query
-
-from Cog import Cog
 
 
 class Quotes(Cog):
